contributor_notifier/pyproject.toml
[tool.poetry]
name = "contributor-notifier"
version = "0.1.0"
description = ""
authors = ["feki-rihab <rihab.feki@fiware.org>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.31.0"
pymongo = "^4.6.1"
pygithub = "^2.2.0"
python-dotenv = "^1.0.1"
github-py = "^0.5.0"
pyyaml = "^6.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
